moving sort code to sort.hpp
